Minor code fixes: final variables, javadoc fixes, Iterator -> for : loops, type casting, primitive unboxing, unnecessary 'else'.

+++ trunk/metafacades/uml/emf/uml22/src/main/java/org/andromda/metafacades/emf/uml22/EntityAssociationLogicImpl.java
@@ -62,7 +62,7 @@ public class EntityAssociationLogicImpl
                             tableNamePrefix,
                             this,
                             UMLProfile.TAGGEDVALUE_PERSISTENCE_TABLE,
-                            ((Entity)type).getMaxSqlNameLength(),
+                            Short.valueOf(((Entity)type).getMaxSqlNameLength()),
                             this.getConfiguredProperty(UMLMetafacadeProperties.SQL_NAME_SEPARATOR));
                 }
             }
@@ -86,8 +86,8 @@ public class EntityAssociationLogicImpl
     }
 
     /**
-     * It is an entity association if both ends are entities (have the entity
-     * stereotype
+     * It is an entity association if both ends are entities (have the entity stereotype)
+     * @return boolean isEntityAssociation
      */
     @Override
     protected boolean handleIsEntityAssociation()
